src/TelegramGPT/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.7.1
    # via httpcore
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
colorama==0.4.6
    # via tqdm
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.27.8
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot==20.4
    # via -r requirements.in
requests==2.31.0
    # via openai
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tqdm==4.65.0
    # via openai
urllib3==2.0.3
    # via requests
yarl==1.9.2
    # via aiohttp
